[cmake] Add support for precompiled headers for Windows
This commit adds a function add_precompiled_header so that precompiled
headers can be added to an existing target.

Unfortunately CMake < 3.1 does not allow to use the same precompiled header
for multiple targets because for debug builds, the pdb files are created in a
per target directory. For precompiled headers to work, we need them in the same
directory, otherwise the following compile error is triggered.

error C2859: <file>.pdb is not the pdb file that was used when this
             precompiled header was created, recreate the precompiled header.

CMake 3.1 resolved this issue and we can specify the path with the
CMAKE_COMPILE_PDB_OUTPUT_DIRECTORY variable. As the function is defined in
the Windows specific scripts directory, this only bumps the CMake minimum
version for Windows to 3.1.

References:
- https://cmake.org/Bug/view.php?id=14667
- https://github.com/Kitware/CMake/commit/fba51b096e2d8ec281653aa05720c11dc9b9bfe6