[cmake] Fix warning CMP0054
CMake >= 3.1 issues a policy warning for statements like
'if("${VAR}" STREQUAL "test")':

CMakeFindBinUtils.cmake:33 (if): Policy CMP0054 is not set: Only interpret if()
arguments as variables or keywords when unquoted. Run "cmake --help-policy
CMP0054" for policy details. Use the cmake_policy command to set the policy and
suppress this warning.

Hence we replace this if statements with 'if(VAR STREQUAL "test")'.