[cmake] Set link_directories for Windows
Windows uses #pragma comment(lib X) in order to specify library
dependencies. To make this code work, we need to specify link_directories
in CMake.

For the future it would certainly be better to use CMake modules to
determine the libraries and to get rid of the #pragma instructions.
Any change here at the moment would break the existing Visual Studio
project, hence we leave this for later.