[cmake] Create variable for configuration specific lib dirs on Windows
Creates CONFIGURATION_LIBDIR, CONFIGURATION_LIBDIR_RELEASE and
CONFIGURATION_LIBDIR_DEBUG variables that can be used in link_directories
and in the FindXXX.cmake modules. Libraries might have the same names for
Release and Debug and can only be distinguished by the directory.

Don't add both directories to link_directories, but only the one for the
matching configuration.