[cmake] Link against precompiled header
Starting with VS2012 the precompiled header object file needs to be linked
into the library/binary using it, otherwise the following error is
triggered: LNK2011: precompiled object not linked in; image may not run

We cannot determine the path to the generated obj file and add it to
the sources using an object library, TARGET_OBJECT and target_sources.
CMake complains about cannot find the obj file.

Instead of an object lib, we create a static lib containing the
precompiled header and link that one.